Use 0 for as refresh key
It makes sense to use 0 as index to URI of currently loaded page.
It's not convenient to type as key for 0 on keyboard is far to the
right but to keep the idea that all numbers are used for selecting
links we can use 0 as index to "first" link being the current page.