Use global tab index value
I didn't liked the previous solution.  This feels better but OFC is
not perfect as I have to manually track the current index.  It's
simple at the moment but might be problematic later.  We will see.
For sure this makes tab structure smaller and logic simpler when new
tab is added but in cost of another global variable.  Time will tell
if this was a good decision, it's good enough for now.