Work on prompt commands system
Single letter commands that can have child commands creating chain of
single letter paths.  This is essentially poor version of spacemacs
and magit interfaces but in CLI style.  When you don't type anything
that reminds command then it is treated as URI.  Otherwise there is an
effort made to match your list of characters to path leading to
action.  If it's not completed meaning that it does not lead to final
action then context menu is printed showing what letters are available
next essentially showing you a sub menu (children) of parent command.

This is a WIP solution.  I will refine it along the way but it already
replaces old implementations.  Plan is to make NULL (empty command) be
the "repeat last" command action.  Also I want to work on visuals
making the context menu a bit more useful.

Now I can also add possibility to display help.  This essentially will
be a context menu of commands for root of all commands.