Refactor, think more in numbers
I thought that definition of protocol value in uri.h lib was too
abstract.  I was already just an number in form of enum but then I
realized that it could be an port number - default port number used by
each protocol.  This opened an path to simpler types in function
parameters and made it easy to handle cases when protocol or port is
missing in URI.