Move Gopher request logic to gph.c
This is a start.  I have to separate as much as possible of request
logic from each protocol.  The more protocols I handle he more
differences I will have.  It's hard to figure out the common part that
can be defined in main.c and right now I struggle with the "show"
logic but I guess I will have to figure it out tomorrow.